CH: Adding base class for zones
