+++ Assets/Scripts/Zones/TimeZone.cs
@@ -1,53 +1,32 @@
-using System.Collections;
 using UnityEngine;
-using UnityEngine.UI;
 
-public class TimeZone : MonoBehaviour
+public class TimeZone : ZoneBase
 {
-    public int id;
-
-    public float triggerTime;
-
-    public GameObject UIPrefab;
-
-    public IntEventChannelSO ZoneEvents;
-
-    private Image ZoneUI;
-    private float timeInside = 0;
-
-    void Awake()
-    {
-        GameObject ZoneUIObject =
-            Instantiate(UIPrefab, transform.position, Quaternion.identity);
-        ZoneUIObject
-            .transform
-            .SetParent(GameObject.Find("SpriteOverlay").transform);
-
-        ZoneUI = ZoneUIObject.GetComponent<Image>();
-    }
-
+    bool activated = false;
     void OnTriggerEnter2D(Collider2D col)
     {
         if (col.gameObject.CompareTag("MothForces"))
         {
-            timeInside = 0;
+            progress = 0;
             ZoneUI.fillAmount = 0;
         }
     }
 
     void OnTriggerStay2D()
     {
-        timeInside += Time.deltaTime;
-        ZoneUI.fillAmount = timeInside / triggerTime;
-        if (timeInside > triggerTime)
+        progress += Time.deltaTime;
+        Mathf.Clamp(progress, 0, activationReq);
+        ZoneUI.fillAmount = progress / activationReq;
+        if (progress >= activationReq && !activated)
         {
             ZoneEvents.RaiseEvent(id);
-            timeInside = 0;
+            activated = true;
         }
     }
 
     void OnTriggerExit2D()
     {
         ZoneUI.fillAmount = 0;
+        activated = false;
     }
 }
