+++ Assets/Scripts/Zones/ZoneBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7b2c6f87917d8efa866cd50d25550f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
